requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.0
flatbuffers==23.5.26
fonttools==4.43.0
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib>=3.4.0
ml-dtypes==0.2.0
numpy>=1.21.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.1
pip==22.0.4
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==58.1.0
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
wheel==0.41.2
wrapt==1.14.1
zipp==3.17.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
pyyaml>=5.4.0
joblib>=1.0.0
tqdm>=4.62.0
pandas>=1.3.0
